requirements.txt
requests
beautifulsoup4
Pillow
PyMuPDF
pytesseract
sqlite-utils
datasette
